# Add default Liquid filter for default values