Instead of

```html
{% if page.title %}
   <title>{{ page.title }}</title>
{% else %}
   <title>Default Title</title>
{% endif %}
```

This filter will allow:
```html
<title>{{ page.title | default: "Default Title" }}</title>
```

This filter is [included in the Liquid 4 release candidate](https://github.com/Shopify/liquid/blob/v4.0.0.rc1/lib/liquid/standardfilters.rb#L344-L347), but I would suspect that it will be a while before Jekyll can move to a new major version of Liquid.

- [x] Add Filter
- [x] Add Tests
- [x] Add Documentation