@parkr 2016-01-12
> I would suspect that it will be a while before Jekyll can move to a new major version of Liquid.

The major release of Liquid may not affect us. We may be able to bump in a 3.2 instead of in a 4.0