@parkr 2016-01-22
@pathawks I'm fine adding this to v3.1.0 if you have time today/tomorrow to add documentation. There's a page for the filters.