@pathawks 2016-01-22
Yes, I will add docs today/tomorrow. :+1: