@pathawks 2016-01-23
Hold on, this looks like it is already [in Liquid 3](https://github.com/Shopify/liquid/blob/v3.0.0/lib/liquid/standardfilters.rb#L286-L289).

Why does it only work if we include it again in Jekyll? Shouldn't it come with [Liquid 3](https://travis-ci.org/jekyll/jekyll/jobs/104242121#L267)