@pathawks 2016-01-23
I feel like this is already included in Liquid and we shouldn't have to duplicate it here, but since the [tests do not pass if we do not include our own `default` filter](https://travis-ci.org/jekyll/jekyll/builds/104242120), I must be mistaken.

Either way, docs have been added and this PR is ready for review by @jekyll/core