@envygeeks 2016-01-23
That code irked me because it was so slow for something so simple (and I know you didn't write it @pathawks) but it's much faster as:

```ruby
def default(str, default = ''.freeze)
  if !str || str == ''.freeze
    return default
  end

  str
end
```

---

#### Ruby 2.3

```
Comparison:
        _new:nil,nil:  3813840.2 i/s
        _old:nil,nil:  2651507.6 i/s - 1.44x slower

Comparison:
      _new:nil,hello:  3173889.5 i/s
      _old:nil,hello:  2301255.8 i/s - 1.38x slower

Comparison:
       _new:'',hello:  2262343.6 i/s
       _old:'',hello:  2119516.2 i/s - 1.07x slower
```

#### Ruby 2.2

```
Comparison:
        _new:nil,nil:  3503265.1 i/s
        _old:nil,nil:  2105375.9 i/s - 1.66x slower

Comparison:
      _new:nil,hello:  2254230.7 i/s
      _old:nil,hello:  1586608.5 i/s - 1.42x slower

Comparison:
       _new:'',hello:  1580059.5 i/s
       _old:'',hello:  1499492.3 i/s - 1.05x slower
```

And you can bump the last two up by another 0.15x if you move `"".freeze` to a constant.