@envygeeks 2016-01-23
I see why they are using empty now, let me adjust my code.