@envygeeks 2016-01-23
Shiny object distracted me, here is a faster version the way it currently is:

```ruby
def _new(obj, default = EMPTY_STR)
  return default unless obj
  if obj.respond_to?(:empty?) && obj.empty?
    default else obj
  end
end
```