@envygeeks 2016-01-23
And the benches if you wish to see them:

```
Comparison:
        _new:nil,nil:  4154265.8 i/s
        _old:nil,nil:  2660080.4 i/s - 1.56x slower

Comparison:
      _new:nil,hello:  3062159.8 i/s
      _old:nil,hello:  2261820.6 i/s - 1.35x slower

Comparison:
       _new:'',hello:  2120534.5 i/s
       _old:'',hello:  2047534.6 i/s - 1.04x slower

Comparison:
       _new:[],hello:  2059903.1 i/s
       _old:[],hello:  2044665.0 i/s - 1.01x slower

Comparison:
       _new:{},hello:  2133132.5 i/s
       _old:{},hello:  2040050.0 i/s - 1.05x slower
```