@parkr 2016-01-24
> I feel like this is already included in Liquid and we shouldn't have to duplicate it here, but since the tests do not pass if we do not include our own default filter, I must be mistaken.

Decided to give this a shot myself and it turns out that it does work! The reason it doesn't work for the tests is because `@filter` refers only to the `JekyllFilters` module, not to all Liquid filters.

```bash
$ echo -e "---\n---\n{{ page.title | default:'Default title' }}" > hi.html
$ jekyll serve
$ open http://localhost:4000/hi.html
# You should see "Default title"
```

I just did this and it works!

<img width="299" alt="screen shot 2016-01-23 at 5 30 18 pm" src="https://cloud.githubusercontent.com/assets/237985/12533896/ff90ea9e-c1f6-11e5-9c58-4867129adbd0.png">

Thanks for putting all the hard work in, @pathawks, and thanks @envygeeks for finding that performance improvement! We'll let the Liquid folks handle it from here.