@pathawks 2016-01-24
> turns out that it does work!

It has been a long week.

Thanks for the sanity check :p